+++ plugins/subagents/runtime/src/index.ts
@@ -1,3 +1,4 @@
+export * from "./memory-budget.js";
 export * from "./model-selection.js";
 export * from "./session-store.js";
 export * from "./subagents.js";